simulation/requirements.txt
mesa
numpy
ape-hardhat
eth-ape
